% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualSHIELDUI.R
\name{VisualSHIELDUI}
\alias{VisualSHIELDUI}
\title{Load main VisualSHIELD UI component
this function generates the UI for the shiny module corresponding to this app.}
\usage{
VisualSHIELDUI(id, title)
}
\arguments{
\item{id}{The id of the module. It should match with the id param of the VisualSHIELDServer function}

\item{title}{The title of the module (as a shiny UI object such as h4) in your custom app.}
}
\description{
Use it like this:
}
\details{
library(shiny)
library(opalr)
library(DSI)
library(DSOpal)
library(dsBaseClient)
library(VisualSHIELD)

shinyUI(
fluidPage(VisualSHIELDUI("VisualSHIELD", h3("Demo VisualSHIELD app title")))
)
}
