% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualSHIELDServer.R
\name{VisualSHIELDServer}
\alias{VisualSHIELDServer}
\title{Load main VisualSHIELD Server component
this function generates the Server counterpart for the UI with the matching id parameter.}
\usage{
VisualSHIELDServer(
  id,
  servers,
  LOG_FILE = "VisualSHIELD.log",
  glm_max_iterations = 30
)
}
\arguments{
\item{id}{The id of the corresponding VisualSHIELD UI.}

\item{servers}{A R list with user information (name/email)and a list of opal/dbNP servers the user can connect to.}

\item{LOG_FILE}{The path to the file where user activity will be logged}

\item{glm_max_iterations}{The maximum number of iterations allowed for the federated analysis}
}
\description{
Load main VisualSHIELD Server component
this function generates the Server counterpart for the UI with the matching id parameter.
}
